src: add color codes for X2
This commit enables color codes for X2. Based on information found in
RattaRleX2Decoder in supernote-tool.

I don't do any parsing of the version field because it seems to not
really matter. If the file has the new color codes just use them. The
supernote-tool codepaths seem to be no-ops besides this.

I do see there is code to support arbitrary grays but I don't see that
being an option in the manual

```
else: # if the color_code is not included in colormap, use the value as color directly
    r, g, b = (color_code, color_code, color_code)
```

Fixes https://github.com/philips/supernote-typescript/issues/2